# Indexing Variable objects with a mask