@fujiisoup 2017-12-01
It is a great idea!
I previously raised #1553 for multidimensional reindex, but am still thinking how could we implement it.
`_getitem_with_mask` looks cleanest.

I will look inside the code later, but I feel this should be exposed to the public.
In this case, what is the most consistent naming with the existing methods?
`isel_mask`? 

> I would probably leave it as a Variable method since this is pretty low-level.

Agreed.