@shoyer 2017-12-04
OK, I'm going to merge this (just the first commit), and leave the second part (actually changing reindex) to another PR.