@fujiisoup 2017-12-04
Could you add more test coverage of the first commit?
It would help other developers to follow the logic.