@mraspaud 2017-12-06
That looks fantastic @shoyer , looking forward to testing it :)