@shoyer 2017-12-13
@fujiisoup could you kindly take another look?