@fujiisoup 2017-12-14
This looks good to me.

I think it is ready to expose `_getitem_with_mask` to public. 
As you mention in #TODO, I also think `isel` is the best place.