@shoyer 2017-12-18
I decided to merge in the current state rather than let this get stale. We can add the public API later....